cmake/prebuilt: Update third-party projects
This is a large refactor that updates the API usage for importing
third-party libraries, now they are discovered within the submodule vs.
downloaded and extracted.

The logic is ported from the old provision code for Homebrew and Choco.
This includes patch files (hence the update to .gitignore) and a
inline-replace python script to handle CMake/sed/perl oddities.

The ncurses library is introduced as a recursive dependency but it is
not used in the final build.